% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crowfly_dist.R
\name{crowfly_dist}
\alias{crowfly_dist}
\title{Crowfly distance between two points}
\usage{
crowfly_dist(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{a numeric value for the longitude of point 1}

\item{lat1}{a numeric value for the latitude of point 1}

\item{long2}{a numeric value for the longitude of point 2}

\item{lat2}{a numeric value for the latitude of point 2}
}
\value{
a numeric value of the distance between point 1 and 2
}
\description{
This functions calculates the crowfly distance between two points (lattitude and logitude).
 The output is in miles.
}
\author{
Thevaa Chandereng
}

