% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_mean.R
\name{new_mean}
\alias{new_mean}
\title{Recurrence Relationship for Mean}
\usage{
new_mean(mu, x, n)
}
\arguments{
\item{mu}{a numeric value of old mean}

\item{x}{a numeric value for the new value to be added to the old mean}

\item{n}{a numeric value for the number of data including the new data}
}
\value{
a numeric value of the updated mean
}
\description{
This functions updates the mean using old mean and new value
}
\details{
This function updates the mean using recurrence relationship between old mean and the new
  value.
}
\author{
Thevaa Chandereng
}
\seealso{
\code{mean}
}

