% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/newvar.R
\name{newvar}
\alias{newvar}
\title{Recurrence Relationship for Sample Variance}
\usage{
newvar(oldvar, n, mu, new)
}
\arguments{
\item{oldvar}{a numeric value of old sample variance}

\item{n}{a numeric value for the number of data including the new data}

\item{mu}{a numeric value of the old mean}

\item{new}{a numeric value of the new data point}
}
\value{
a numeric value of the updated sample variance
}
\description{
This functions updates the sample variance using old sample variance, old mean and new value
}
\details{
This function updates the sample variance using recurrence relationship between old mean, old sample variance and the new
  value.
}
\author{
Thevaa Chandereng
}
\seealso{
\code{var}

\code{new_mean}
}

