% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratio.R
\name{oddsratio}
\alias{oddsratio}
\title{Odds ratio for 2x2 table with Confidence Interval}
\usage{
oddsratio(x, CI = 0.95)
}
\description{
This functions prints a list of odds ratio and confidence interval for a 2x2
 count data
}
\details{
@param x a vector of length 4 with the data in the order of a, b, c, d
  @param CI a numeric vector indicating the two sided confidence interval
  @return a list of two elements: element 1 is the odds ratio, element 2 is a vector of length
  two of lower bound and upper bounds of the odds ratio
  @author Thevaa Chandereng
  @seealso \code{mean}
  @export
}

